Split synced rollback regression test
